breaking apart init logic
