+++ pkg/skaffold/initializer/init.go
@@ -19,32 +19,23 @@ package initializer
 import (
 	"bufio"
 	"context"
-	"encoding/json"
 	"fmt"
 	"io"
 	"io/ioutil"
 	"os"
-	"os/exec"
 	"path/filepath"
-	"regexp"
 	"sort"
 	"strings"
 
-	"github.com/karrick/godirwalk"
-	"github.com/pkg/errors"
-	"github.com/sirupsen/logrus"
-	"gopkg.in/AlecAivazis/survey.v1"
-	"gopkg.in/yaml.v2"
-
 	"github.com/GoogleContainerTools/skaffold/cmd/skaffold/app/tips"
-	"github.com/GoogleContainerTools/skaffold/pkg/skaffold/build/buildpacks"
-	"github.com/GoogleContainerTools/skaffold/pkg/skaffold/build/jib"
 	"github.com/GoogleContainerTools/skaffold/pkg/skaffold/config"
 	"github.com/GoogleContainerTools/skaffold/pkg/skaffold/docker"
 	"github.com/GoogleContainerTools/skaffold/pkg/skaffold/initializer/kubectl"
 	"github.com/GoogleContainerTools/skaffold/pkg/skaffold/schema/latest"
 	"github.com/GoogleContainerTools/skaffold/pkg/skaffold/util"
 	"github.com/GoogleContainerTools/skaffold/pkg/skaffold/warnings"
+	"github.com/karrick/godirwalk"
+	"github.com/pkg/errors"
 )
 
 // For testing
@@ -226,390 +217,6 @@ func DoInit(ctx context.Context, out io.Writer, c Config) error {
 	return nil
 }
 
-// runKompose runs the `kompose` CLI before running skaffold init
-func runKompose(ctx context.Context, composeFile string) error {
-	if _, err := os.Stat(composeFile); os.IsNotExist(err) {
-		return err
-	}
-
-	logrus.Infof("running 'kompose convert' for file %s", composeFile)
-	komposeCmd := exec.CommandContext(ctx, "kompose", "convert", "-f", composeFile)
-	_, err := util.RunCmdOut(komposeCmd)
-	return err
-}
-
-// autoSelectBuilders takes a list of builders and images, checks if any of the builders' configured target
-// images match an image in the image list, and returns a list of the matching builder/image pairs. Also
-// separately returns the builder configs and images that didn't have any matches.
-func autoSelectBuilders(builderConfigs []InitBuilder, images []string) ([]builderImagePair, []InitBuilder, []string) {
-	var pairs []builderImagePair
-	var unresolvedImages = make(set)
-	for _, image := range images {
-		matchingConfigIndex := -1
-		for i, config := range builderConfigs {
-			if image != config.ConfiguredImage() {
-				continue
-			}
-
-			// Found more than one match; can't auto-select.
-			if matchingConfigIndex != -1 {
-				matchingConfigIndex = -1
-				break
-			}
-			matchingConfigIndex = i
-		}
-
-		if matchingConfigIndex != -1 {
-			// Exactly one pair found; save the pair and remove from remaining build configs
-			pairs = append(pairs, builderImagePair{ImageName: image, Builder: builderConfigs[matchingConfigIndex]})
-			builderConfigs = append(builderConfigs[:matchingConfigIndex], builderConfigs[matchingConfigIndex+1:]...)
-		} else {
-			// No definite pair found, add to images list
-			unresolvedImages.add(image)
-		}
-	}
-	return pairs, builderConfigs, unresolvedImages.values()
-}
-
-// detectBuilders checks if a path is a builder config, and if it is, returns the InitBuilders representing the
-// configs. Also returns a boolean marking search completion for subdirectories (true = subdirectories should
-// continue to be searched, false = subdirectories should not be searched for more builders)
-func detectBuilders(enableJibInit, enableBuildpackInit bool, path string) ([]InitBuilder, bool) {
-	// TODO: Remove backwards compatibility if statement (not entire block)
-	if enableJibInit {
-		// Check for jib
-		if builders := jib.Validate(path); builders != nil {
-			results := make([]InitBuilder, len(builders))
-			for i := range builders {
-				results[i] = builders[i]
-			}
-			return results, false
-		}
-	}
-
-	// Check for Dockerfile
-	base := filepath.Base(path)
-	if strings.Contains(strings.ToLower(base), "dockerfile") {
-		if docker.Validate(path) {
-			results := []InitBuilder{docker.ArtifactConfig{File: path}}
-			return results, true
-		}
-	}
-
-	// TODO: Remove backwards compatibility if statement (not entire block)
-	if enableBuildpackInit {
-		// Check for buildpacks
-		if buildpacks.Validate(path) {
-			results := []InitBuilder{buildpacks.ArtifactConfig{File: path}}
-			return results, true
-		}
-	}
-
-	// TODO: Check for more builders
-
-	return nil, true
-}
-
-func processCliArtifacts(artifacts []string) ([]builderImagePair, error) {
-	var pairs []builderImagePair
-	for _, artifact := range artifacts {
-		// Parses JSON in the form of: {"builder":"Name of Builder","payload":{...},"image":"image.name"}.
-		// The builder field is parsed first to determine the builder type, and the payload is parsed
-		// afterwards once the type is determined.
-		a := struct {
-			Name  string `json:"builder"`
-			Image string `json:"image"`
-		}{}
-		if err := json.Unmarshal([]byte(artifact), &a); err != nil {
-			// Not JSON, use backwards compatible method
-			parts := strings.Split(artifact, "=")
-			if len(parts) != 2 {
-				return nil, fmt.Errorf("malformed artifact provided: %s", artifact)
-			}
-			pairs = append(pairs, builderImagePair{
-				Builder:   docker.ArtifactConfig{File: parts[0]},
-				ImageName: parts[1],
-			})
-			continue
-		}
-
-		// Use builder type to parse payload
-		switch a.Name {
-		case docker.Name:
-			parsed := struct {
-				Payload docker.ArtifactConfig `json:"payload"`
-			}{}
-			if err := json.Unmarshal([]byte(artifact), &parsed); err != nil {
-				return nil, err
-			}
-			pair := builderImagePair{Builder: parsed.Payload, ImageName: a.Image}
-			pairs = append(pairs, pair)
-
-		// FIXME: shouldn't use a human-readable name?
-		case jib.PluginName(jib.JibGradle), jib.PluginName(jib.JibMaven):
-			parsed := struct {
-				Payload jib.ArtifactConfig `json:"payload"`
-			}{}
-			if err := json.Unmarshal([]byte(artifact), &parsed); err != nil {
-				return nil, err
-			}
-			parsed.Payload.BuilderName = a.Name
-			pair := builderImagePair{Builder: parsed.Payload, ImageName: a.Image}
-			pairs = append(pairs, pair)
-
-		case buildpacks.Name:
-			parsed := struct {
-				Payload buildpacks.ArtifactConfig `json:"payload"`
-			}{}
-			if err := json.Unmarshal([]byte(artifact), &parsed); err != nil {
-				return nil, err
-			}
-			pair := builderImagePair{Builder: parsed.Payload, ImageName: a.Image}
-			pairs = append(pairs, pair)
-
-		default:
-			return nil, errors.New("unknown builder type in CLI artifacts")
-		}
-	}
-	return pairs, nil
-}
-
-// For each image parsed from all k8s manifests, prompt the user for the builder that builds the referenced image
-func resolveBuilderImages(builderConfigs []InitBuilder, images []string, force bool) ([]builderImagePair, error) {
-	// If nothing to choose, don't bother prompting
-	if len(images) == 0 || len(builderConfigs) == 0 {
-		return []builderImagePair{}, nil
-	}
-
-	// if we only have 1 image and 1 build config, don't bother prompting
-	if len(images) == 1 && len(builderConfigs) == 1 {
-		return []builderImagePair{{
-			Builder:   builderConfigs[0],
-			ImageName: images[0],
-		}}, nil
-	}
-
-	if force {
-		return nil, errors.New("unable to automatically resolve builder/image pairs; run `skaffold init` without `--force` to manually resolve ambiguities")
-	}
-
-	// Build map from choice string to builder config struct
-	choices := make([]string, len(builderConfigs))
-	choiceMap := make(map[string]InitBuilder, len(builderConfigs))
-	for i, buildConfig := range builderConfigs {
-		choice := buildConfig.Describe()
-		choices[i] = choice
-		choiceMap[choice] = buildConfig
-	}
-	sort.Strings(choices)
-
-	// For each choice, use prompt string to pair builder config with k8s image
-	pairs := []builderImagePair{}
-	for {
-		if len(images) == 0 {
-			break
-		}
-
-		image := images[0]
-		choice, err := promptUserForBuildConfigFunc(image, choices)
-		if err != nil {
-			return nil, err
-		}
-
-		if choice != NoBuilder {
-			pairs = append(pairs, builderImagePair{Builder: choiceMap[choice], ImageName: image})
-			choices = util.RemoveFromSlice(choices, choice)
-		}
-		images = util.RemoveFromSlice(images, image)
-	}
-	if len(choices) > 0 {
-		logrus.Warnf("unused builder configs found in repository: %v", choices)
-	}
-	return pairs, nil
-}
-
-func promptUserForBuildConfig(image string, choices []string) (string, error) {
-	var selectedBuildConfig string
-	options := append(choices, NoBuilder)
-	prompt := &survey.Select{
-		Message:  fmt.Sprintf("Choose the builder to build image %s", image),
-		Options:  options,
-		PageSize: 15,
-	}
-	err := survey.AskOne(prompt, &selectedBuildConfig, nil)
-	if err != nil {
-		return "", err
-	}
-
-	return selectedBuildConfig, nil
-}
-
-func artifacts(pairs []builderImagePair) []*latest.Artifact {
-	var artifacts []*latest.Artifact
-
-	for _, pair := range pairs {
-		artifact := &latest.Artifact{
-			ImageName: pair.ImageName,
-		}
-
-		workspace := filepath.Dir(pair.Builder.Path())
-		if workspace != "." {
-			artifact.Workspace = workspace
-		}
-
-		pair.Builder.UpdateArtifact(artifact)
-
-		artifacts = append(artifacts, artifact)
-	}
-
-	return artifacts
-}
-
-func generateSkaffoldConfig(k Initializer, buildConfigPairs []builderImagePair) ([]byte, error) {
-	// if we're here, the user has no skaffold yaml so we need to generate one
-	// if the user doesn't have any k8s yamls, generate one for each dockerfile
-	logrus.Info("generating skaffold config")
-
-	name, err := suggestConfigName()
-	if err != nil {
-		warnings.Printf("Couldn't generate default config name: %s", err.Error())
-	}
-
-	return yaml.Marshal(&latest.SkaffoldConfig{
-		APIVersion: latest.Version,
-		Kind:       "Config",
-		Metadata: latest.Metadata{
-			Name: name,
-		},
-		Pipeline: latest.Pipeline{
-			Build: latest.BuildConfig{
-				Artifacts: artifacts(buildConfigPairs),
-			},
-			Deploy: k.GenerateDeployConfig(),
-		},
-	})
-}
-
-func suggestConfigName() (string, error) {
-	cwd, err := os.Getwd()
-	if err != nil {
-		return "", err
-	}
-
-	base := filepath.Base(cwd)
-
-	// give up for edge cases
-	if base == "." || base == string(filepath.Separator) {
-		return "", nil
-	}
-
-	return canonicalizeName(base), nil
-}
-
-// canonicalizeName converts a given string to a valid k8s name string.
-// See https://kubernetes.io/docs/concepts/overview/working-with-objects/names/#names for details
-func canonicalizeName(name string) string {
-	forbidden := regexp.MustCompile(`[^-.a-z]+`)
-	canonicalized := forbidden.ReplaceAllString(strings.ToLower(name), "-")
-	if len(canonicalized) <= 253 {
-		return canonicalized
-	}
-	return canonicalized[:253]
-}
-
-func printAnalyzeJSONNoJib(out io.Writer, skipBuild bool, pairs []builderImagePair, unresolvedBuilders []InitBuilder, unresolvedImages []string) error {
-	if !skipBuild && len(unresolvedBuilders) == 0 {
-		return errors.New("one or more valid Dockerfiles must be present to build images with skaffold; please provide at least one Dockerfile and try again, or run `skaffold init --skip-build`")
-	}
-
-	a := struct {
-		Dockerfiles []string `json:"dockerfiles,omitempty"`
-		Images      []string `json:"images,omitempty"`
-	}{Images: unresolvedImages}
-
-	for _, pair := range pairs {
-		if pair.Builder.Name() == docker.Name {
-			a.Dockerfiles = append(a.Dockerfiles, pair.Builder.Path())
-		}
-		a.Images = append(a.Images, pair.ImageName)
-	}
-	for _, config := range unresolvedBuilders {
-		if config.Name() == docker.Name {
-			a.Dockerfiles = append(a.Dockerfiles, config.Path())
-		}
-	}
-
-	return printJSON(out, a)
-}
-
-// printAnalyzeJSON takes the automatically resolved builder/image pairs, the unresolved images, and the unresolved builders, and generates
-// a JSON string containing builder config information,
-func printAnalyzeJSON(out io.Writer, skipBuild bool, pairs []builderImagePair, unresolvedBuilders []InitBuilder, unresolvedImages []string) error {
-	if !skipBuild && len(unresolvedBuilders) == 0 {
-		return errors.New("one or more valid Dockerfiles must be present to build images with skaffold; please provide at least one Dockerfile and try again, or run `skaffold init --skip-build`")
-	}
-
-	// Build JSON output. Example schema is below:
-	// {
-	//     "builders":[
-	//         {
-	//             "name":"Docker",
-	//             "payload":"path/to/Dockerfile"
-	//         },
-	//         {
-	//             "name":"Name of Builder",
-	//             "payload": { // Payload structure may vary depending on builder type
-	//                 "path":"path/to/builder.config",
-	//                 "targetImage":"gcr.io/project/images",
-	//                 ...
-	//             }
-	//         },
-	//     ],
-	//     "images":[
-	//         {"name":"gcr.io/project/images", "foundMatch":"true"}, // No need to prompt for this image since its builder was automatically resolved
-	//         {"name":"another/image", "foundMatch":"false"},
-	//     ],
-	// }
-	//
-	// "builders" is the list of builder configurations, and contains a builder name and a builder-specific payload
-	// "images" contains an image name and a boolean that indicates whether a builder/image pair can be automatically resolved (true) or if it requires prompting (false)
-	type Builder struct {
-		Name    string      `json:"name,omitempty"`
-		Payload InitBuilder `json:"payload"`
-	}
-	type Image struct {
-		Name       string `json:"name"`
-		FoundMatch bool   `json:"foundMatch"`
-	}
-	a := struct {
-		Builders []Builder `json:"builders,omitempty"`
-		Images   []Image   `json:"images,omitempty"`
-	}{}
-
-	for _, pair := range pairs {
-		a.Builders = append(a.Builders, Builder{Name: pair.Builder.Name(), Payload: pair.Builder})
-		a.Images = append(a.Images, Image{Name: pair.ImageName, FoundMatch: true})
-	}
-	for _, config := range unresolvedBuilders {
-		a.Builders = append(a.Builders, Builder{Name: config.Name(), Payload: config})
-	}
-	for _, image := range unresolvedImages {
-		a.Images = append(a.Images, Image{Name: image, FoundMatch: false})
-	}
-
-	return printJSON(out, a)
-}
-
-func printJSON(out io.Writer, v interface{}) error {
-	contents, err := json.Marshal(v)
-	if err != nil {
-		return errors.Wrap(err, "marshalling contents")
-	}
-
-	_, err = out.Write(contents)
-	return err
-}
-
 // walk recursively walks a directory and returns the k8s configs and builder configs that it finds
 func walk(dir string, force, enableJibInit, enableBuildpackInit bool) ([]string, []InitBuilder, error) {
 	var potentialConfigs []string
@@ -668,23 +275,4 @@ func walk(dir string, force, enableJibInit, enableBuildpackInit bool) ([]string,
 		return nil, nil, err
 	}
 	return potentialConfigs, foundBuilders, nil
-}
-
-// checkConfigFile checks if filePath is a skaffold config or k8s config, or builder config. Detected k8s configs are added to potentialConfigs.
-// Returns true if filePath is a config file, and false if not.
-func checkConfigFile(filePath string, force bool, potentialConfigs *[]string) (bool, error) {
-	if IsSkaffoldConfig(filePath) {
-		if !force {
-			return true, fmt.Errorf("pre-existing %s found", filePath)
-		}
-		logrus.Debugf("%s is a valid skaffold configuration: continuing since --force=true", filePath)
-		return true, nil
-	}
-
-	if kubectl.IsKubernetesManifest(filePath) {
-		*potentialConfigs = append(*potentialConfigs, filePath)
-		return true, nil
-	}
-
-	return false, nil
 }
